Create fscraper-auto-v5
This program scrapes the website mystore411.com, generating data to be uploaded into a db.

This is a debugged version of v4 code that crashed when the function for finding the place_id had an error. As this program is now only generates sample data for verification purposes, and data stills need to be formatted properly.