Create fscraper-auto-dummy
This program extracts data from the website mystore411.com and generates the output without performing queries using Google Maps API, just to get an estimate of the volume of data that will be generated. For now it errors out when reading cities URL's and it is what I am investigating right now.