find-nearest-ws.py
This program finds the nearest weather station according to their coordinates to a given postal code supplied by the user in format "MXM MXM".  The program looks for the weather stations in the "weather-stations" table of the database "weather" residing in the VM, according to the first letter of the postal code entered, locates all the weather stations in the table that matches that letter. Then use Geopy to calculate the distance between the postal code entered and each postal code found in the database (using the coordinates of the postal codes).